AMP-457 M&E feature
git-svn-id: file:///svn/trunk@465 9e8d75aa-d0cb-44a9-8341-a4b8ba1a38b4

+++ digijava/WEB-INF/src/org/digijava/module/aim/dbentity/AmpIndicatorRiskRatings.java
@@ -1,14 +1,14 @@
 package org.digijava.module.aim.dbentity;
 
 public class AmpIndicatorRiskRatings {
-	private Integer ampIndRiskRatingsId;
+	private Long ampIndRiskRatingsId;
 	private String ratingName;
 	private int ratingValue;
 	
-	public Integer getAmpIndRiskRatingsId() {
+	public Long getAmpIndRiskRatingsId() {
 		return ampIndRiskRatingsId;
 	}
-	public void setAmpIndRiskRatingsId(Integer ampIndRiskRatingsId) {
+	public void setAmpIndRiskRatingsId(Long ampIndRiskRatingsId) {
 		this.ampIndRiskRatingsId = ampIndRiskRatingsId;
 	}
 	public String getRatingName() {
